JAMES-3492, move package, use backend-es-v7 packages in imports
