+++ mailbox/plugin/quota-search-elasticsearch-v7/src/main/java/org/apache/james/quota/search/elasticsearch/v7/events/ElasticSearchQuotaMailboxListener.java
@@ -16,20 +16,20 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-package org.apache.james.quota.search.elasticsearch.events;
+package org.apache.james.quota.search.elasticsearch.v7.events;
 
 import javax.inject.Inject;
 import javax.inject.Named;
 
-import org.apache.james.backends.es.DocumentId;
-import org.apache.james.backends.es.ElasticSearchIndexer;
-import org.apache.james.backends.es.RoutingKey;
+import org.apache.james.backends.es.v7.DocumentId;
+import org.apache.james.backends.es.v7.ElasticSearchIndexer;
+import org.apache.james.backends.es.v7.RoutingKey;
 import org.apache.james.core.Username;
 import org.apache.james.mailbox.events.Event;
 import org.apache.james.mailbox.events.Group;
 import org.apache.james.mailbox.events.MailboxListener;
-import org.apache.james.quota.search.elasticsearch.QuotaRatioElasticSearchConstants;
-import org.apache.james.quota.search.elasticsearch.json.QuotaRatioToElasticSearchJson;
+import org.apache.james.quota.search.elasticsearch.v7.QuotaRatioElasticSearchConstants;
+import org.apache.james.quota.search.elasticsearch.v7.json.QuotaRatioToElasticSearchJson;
 import org.reactivestreams.Publisher;
 
 import reactor.core.publisher.Mono;
